Refactor: apply re-export
